parse int to byte

+++ FirstApp/Program.cs
@@ -10,7 +10,7 @@ namespace FirstApp
             string name = Console.ReadLine();
 
             Console.Write("Enter your age: ");
-            int age= int.Parse(Console.ReadLine());
+            byte age= (byte)int.Parse(Console.ReadLine());
 
             Console.WriteLine("Your name is {0} and age is {1} ", name, age);
